remove unnecessary model initialization
Formerly, two quite large queries for work packages where issued and the
results of both where initialized as AR models.

The first query was issued to get the work packages on which most of the
collection representer's calculations where performend.

The second query was issued to represent the individual work packages in
the _elements list. Here, eager loading was performed.

The PR moves the eager loading to the first query and alters the
representer flow so that the results of the first query is also employed
for the individual work package representers.

That way, the necessary AR models do only have to be instantiated once
which saves quite some time on a request.